refactor: app info
